Extract build receipt into own jar
Instead of including the build receipt into
the core jar we add it to an own jar - version-info.
The build receipt often contains a timestamp and therefore
we want to use it as little as possible.
By doing this we keep the classpath constant between
builds and therefore allow for more caching.

We also have to add the version-info to the Gradle Worker process.
This makes it possible to evaluate the `GradleVersion`
in those, too. If one wants to log deprecation messages
from there this is necessary. For example`JULRedirectorIntegrationTest.defaultLoggingConfigNoFineLevelWhenDisabled`
fails without this.