Remove wait period before measurements
The wait period gives the JVM a pause to do a full GC. This means
that the measurements afterwards will start out fast and then get
slower again as the heap fills up. Instead, we now go directly from
warmup to measurement, with the heap in a more realistic "load" state.