Validate tasks executed when using typed configuration views
Note that at this stage all artifacts for a configuration will be built
for a configuration, even if only some of those artifacts are required
for a configuration view. Fixing this isn't really part of the 
`DependencyTransform` work, but will be important to avoid building the
AAR file in order to compile java sources that reference an Android
library.